Factored out sections in overview tab into subcomponents.
